#10244: Move some utils/test infra for resnet
