+++ tests/ttnn/integration_tests/resnet/test_performance.py
@@ -6,7 +6,7 @@ from loguru import logger
 import time
 import pytest
 
-from tests.ttnn.integration_tests.resnet.test_ttnn_functional_resnet50 import create_test_infra
+from models.demos.ttnn_resnet.tests.ttnn_resnet_test_infra import create_test_infra
 from models.utility_functions import skip_for_wormhole_b0, skip_for_grayskull
 import ttnn
 from models.perf.device_perf_utils import run_device_perf, check_device_perf, prep_device_perf_report
